requirements.txt
Django==1.9.7
Markdown==2.6.6
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
Twisted-Core==13.2.0
Twisted-Web==13.2.0
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
argparse==1.2.1
betterpath==0.2.2
chardet==2.0.1
click==6.6
colorama==0.2.5
command-not-found==0.3
debtagshw==0.1
defer==1.0.6
dirspec==13.10
django-autoslug==1.9.3
django-ckeditor==5.0.3
django-cleanup==0.4.2
django-contrib-comments==1.7.1
django-markdown==0.8.4
dnspython==1.11.1
duplicity==0.6.23
get==0.0.10
grabber==0.1.0
html5lib==0.999
httplib2==0.8
l==0.11.0
lockfile==0.8
lxml==3.3.3
oauthlib==0.6.1
oneconf==0.3.7
pexpect==3.1
piston-mini-client==0.7.5
post==0.0.7
public==0.0.32
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5
python-debian==0.1.21-nmu2ubuntu2
pyxdg==0.25
query-string==0.0.7
reportlab==3.0
request==0.0.7
requests==2.2.1
sessioninstaller==0.0.0
six==1.5.2
software-center-aptd-plugins==0.0.0
ssh-import-id==3.21
system-service==0.1.6
unity-lens-photos==1.0
urlgrabber==3.9.1
urllib3==1.7.1
vcversioner==2.16.0.0
virtualenv==15.0.1
wheel==0.24.0
whitenoise==3.2
wsgiref==0.1.2
xdiagnose==3.6.3build2
zope.interface==4.0.5
